Move workflow history view
