+++ arches/app/utils/request.py
@@ -1,7 +1,6 @@
 def looks_like_api_call(request):
     known_api_paths_without_api_signposting = [
         # Until these are renamed for consistency, just list them.
-        "arches.app.views.workflow_history",
     ]
     if "/api/" in request.path:
         return True
