+++ releases/8.0.0.md
@@ -135,7 +135,7 @@ JavaScript:
 
 - `GraphModel.get_published_graph()` no longer accepts the `raise_if_missing` argument.
 - `Graph.new()` has been deprecated in favor of `Graph.objects.create_graph()`
-
+- `WorkflowHistoryView` moved from `arches.app.views.workflow_history` to `arches.app.views.api`.
 - The format of JavaScript files has changed. Please refer to [Upgrading an Arches project](#upgrading-an-arches-project).
 - The format of how some images are loaded in template files has changed. Please refer to [Upgrading an Arches project](#upgrading-an-arches-project).
 - Update any imports from `@primevue/themes` to `@primeuix/themes`.
